requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn==0.32.1
python-telegram-bot==21.8
openai==1.99.7
python-dotenv==1.0.1

# Google Sheets integration (optional)
gspread==6.1.4
google-auth==2.36.0

# YouTube search
requests==2.32.3

# Basic utilities
pydantic==2.10.2
